+++ solanaceae/message3/file_rw_file.hpp
@@ -29,7 +29,7 @@ struct FileRWFile : public FileI {
 		return _file.is_open() && _file.good();
 	}
 
-	std::vector<uint8_t> read(uint64_t pos, uint32_t size) override {
+	std::vector<uint8_t> read(uint64_t pos, uint64_t size) override {
 		if (pos >= _file_size) {
 			return {};
 		}
